Required path for article render
