+++ frictionless/server/helpers/__init__.py
@@ -1,10 +1,8 @@
-from .article import *
 from .file import *
 from .general import *
 from .json import *
 from .measure import *
 from .record import *
 from .resource import *
-from .script import *
 from .table import *
 from .text import *
